+++ kinesis-consumer.js
@@ -26,7 +26,7 @@ const streamProcessing = require('aws-stream-consumer-core/stream-processing');
  * @module kinesis-stream-consumer/kinesis-consumer
  * @author Byron du Preez
  */
-exports._ = '_'; //IDE workaround
+exports._$_ = '_$_'; //IDE workaround
 
 // Configuration
 exports.isStreamConsumerConfigured = isStreamConsumerConfigured;
@@ -94,12 +94,13 @@ function configureStreamConsumer(context, settings, options, event, awsContext) 
   return context;
 }
 
+ // * @param {undefined|ConfigureStreamConsumer} [configureStreamConsumer] - an optional `configureStreamConsumer` function to be used instead of the default `configureKinesisStreamConsumer` function
 /**
  * Generates a handler function for your Kinesis stream consumer Lambda.
  *
  * @param {undefined|function(): (Object|StreamConsumerContext|StreamProcessing|StandardContext)} createContext - a optional function that will be used to create the initial context to be configured & used
- * @param {undefined|StreamConsumerSettings|function(): StreamConsumerSettings} [generateSettings] - an optional function that will be used to generate initial stream consumer settings to use; OR optional module-scoped stream consumer settings from which to copy initial stream consumer settings to use
- * @param {undefined|StreamConsumerOptions|function(): StreamConsumerOptions} [generateOptions] - an optional function that will be used to generate initial stream consumer options to use; OR optional module-scoped stream consumer options from which to copy initial stream consumer options to use
+ * @param {undefined|StreamConsumerSettings|function(): StreamConsumerSettings} [createSettings] - an optional function that will be used to generate initial stream consumer settings to use; OR optional module-scoped stream consumer settings from which to copy initial stream consumer settings to use
+ * @param {undefined|StreamConsumerOptions|function(): StreamConsumerOptions} [createOptions] - an optional function that will be used to generate initial stream consumer options to use; OR optional module-scoped stream consumer options from which to copy initial stream consumer options to use
  * @param {undefined|function(): TaskDef[]} [defineProcessOneTasks] - an "optional" function that must generate a new list of "processOne" task definitions, which will be subsequently used to generate the tasks to be executed on each message independently
  * @param {undefined|function(): TaskDef[]} [defineProcessAllTasks] - an "optional" function that must generate a new list of "processAll" task definitions, which will be subsequently used to generate the tasks to be executed on all of the event's messages collectively
  * @param {Object|undefined} [opts] - optional options to use to configure the generated handler function
@@ -108,7 +109,7 @@ function configureStreamConsumer(context, settings, options, event, awsContext) 
  * @param {string|undefined} [opts.successMsg] an optional message to log at info level on success
  * @returns {AwsLambdaHandlerFunction} a handler function for your stream consumer Lambda
  */
-function generateHandlerFunction(createContext, generateSettings, generateOptions, defineProcessOneTasks,
+function generateHandlerFunction(createContext, createSettings, createOptions, defineProcessOneTasks,
   defineProcessAllTasks, opts) {
 
   const logEventResultAtLogLevel = opts && logging.isValidLogLevel(opts.logEventResultAtLogLevel) ?
@@ -128,11 +129,11 @@ function generateHandlerFunction(createContext, generateSettings, generateOption
       // Configure the context as a stream consumer context
       context = typeof createContext === 'function' ? createContext() : {};
 
-      const settings = typeof generateSettings === 'function' ? copy(generateSettings(), deep) :
-        generateSettings && typeof generateSettings === 'object' ? copy(generateSettings, deep) : undefined;
+      const settings = typeof createSettings === 'function' ? copy(createSettings(), deep) :
+        createSettings && typeof createSettings === 'object' ? copy(createSettings, deep) : undefined;
 
-      const options = typeof generateOptions === 'function' ? copy(generateOptions(), deep) :
-        generateOptions && typeof generateOptions === 'object' ? copy(generateOptions, deep) : undefined;
+      const options = typeof createOptions === 'function' ? copy(createOptions(), deep) :
+        createOptions && typeof createOptions === 'object' ? copy(createOptions, deep) : undefined;
 
       context = configureStreamConsumer(context, settings, options, event, awsContext);